Refactor: separate deprecated syscall handler in different files
