+++ src/business_logic/execution/execution_entry_point.rs
@@ -5,8 +5,8 @@ use crate::{
         state::state_api::StateReader, transaction::error::TransactionError,
     },
     core::syscalls::{
-        business_logic_syscall_handler::DeprecatedBLSyscallHandler,
-        syscall_handler::{DeprecatedSyscallHandler, SyscallHintProcessor},
+        deprecated_business_logic_syscall_handler::DeprecatedBLSyscallHandler,
+        deprecated_syscall_handler::{DeprecatedSyscallHandler, SyscallHintProcessor},
     },
     definitions::{constants::DEFAULT_ENTRY_POINT_SELECTOR, general_config::StarknetGeneralConfig},
     services::api::contract_class::{ContractClass, ContractEntryPoint, EntryPointType},
