+++ src/core/syscalls/mod.rs
@@ -1,4 +1,6 @@
 pub mod business_logic_syscall_handler;
+pub mod deprecated_business_logic_syscall_handler;
+pub mod deprecated_syscall_handler;
 pub mod deprecated_syscall_request;
 pub mod deprecated_syscall_response;
 pub mod hint_code;
